Move GSheets upload logic for reuse between AsyncTask and Worker
Removes possibility to cancel an upload between rows. Ideally, the upload of each form would be atomic so this is an improvement. It also better matches the server upload.
